fix(frontend): several history improvements
